RSE-32: Refactor
